// admin/src/modulos/api/model/medicamento.ts
/**
 * Receita Digital
 * Definição da API da Receita Digital 
 *
 * OpenAPI spec version: 1.0.6
 * Contact: fran.fig@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { MedicamentoApresentacao } from './medicamentoApresentacao';
import { MedicamentoClasse } from './medicamentoClasse';
import { MedicamentoLaboratorio } from './medicamentoLaboratorio';
import { MedicamentoMedicamentoReferencia } from './medicamentoMedicamentoReferencia';
import { MedicamentoReceita } from './medicamentoReceita';
import { Tag } from './tag';


export interface Medicamento {
    /**
     * código global único da medicamento, usado no
     */
    id?: string;
    /**
     * classificação de risco do medicamento
     */
    tarja?: Medicamento.TarjaEnum;
    receita?: MedicamentoReceita;
    /**
     * Nome oficial do produto
     */
    nome?: string;
    /**
     * Princípio Ativo (se separado por ; ele vai estar disciminado no campo principiosAtivos)
     */
    principio?: string;
    principiosAtivos?: Array<string>;
    apresentacao?: Array<MedicamentoApresentacao>;
    classe?: MedicamentoClasse;
    /**
     * Tipo de produto
     */
    tipo?: Medicamento.TipoEnum;
    laboratorio?: MedicamentoLaboratorio;
    /**
     * O rémedio é de uso restrito hospitalar
     */
    restricaoHospitalar?: boolean;
    medicamentoReferencia?: MedicamentoMedicamentoReferencia;
    tags?: Array<Tag>;
}
export namespace Medicamento {
    export type TarjaEnum = 'Tarja Preta' | 'Tarja Vermelha' | 'Venda Livre';
    export const TarjaEnum = {
        TarjaPreta: 'Tarja Preta' as TarjaEnum,
        TarjaVermelha: 'Tarja Vermelha' as TarjaEnum,
        VendaLivre: 'Venda Livre' as TarjaEnum
    }
    export type TipoEnum = 'Biológico Novo' | 'Biológicos' | 'Específico' | 'Específico (Referência)' | 'Genérico' | 'Genérico (Referência)' | 'Novo' | 'Novo (Referência)' | 'Radiofármaco' | 'Similar' | 'Similar (Referência)';
    export const TipoEnum = {
        BiolgicoNovo: 'Biológico Novo' as TipoEnum,
        Biolgicos: 'Biológicos' as TipoEnum,
        Especfico: 'Específico' as TipoEnum,
        EspecficoReferncia: 'Específico (Referência)' as TipoEnum,
        Genrico: 'Genérico' as TipoEnum,
        GenricoReferncia: 'Genérico (Referência)' as TipoEnum,
        Novo: 'Novo' as TipoEnum,
        NovoReferncia: 'Novo (Referência)' as TipoEnum,
        Radiofrmaco: 'Radiofármaco' as TipoEnum,
        Similar: 'Similar' as TipoEnum,
        SimilarReferncia: 'Similar (Referência)' as TipoEnum
    }
}