// admin/src/modulos/api/model/medicamentoApresentacao.ts
/**
 * Receita Digital
 * Definição da API da Receita Digital 
 *
 * OpenAPI spec version: 1.0.6
 * Contact: fran.fig@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { MedicamentoPrecos } from './medicamentoPrecos';


export interface MedicamentoApresentacao {
    /**
     * código ean do produto/médicamento (código de barras)
     */
    ean?: string;
    /**
     * registro na anvisa do produto/médicamento
     */
    registro?: string;
    /**
     * registro específico da anvisa do produto/médicamento
     */
    ggerem?: string;
    /**
     * Descrevem a apresentação do produto
     */
    descricao?: string;
    precos?: Array<MedicamentoPrecos>;
}